"refactor: move pagination to common component and improve pagination ui"
